Refactor `/github` endpoint into it's own file
